+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - USGS disclaimer & FAQ
 - User's Guide
 - Metadata section with CSV downloads in user's guide
+- Stores flag data in local storage which gives user option to use them after closing tab
+- Browser warning when refreshing or closing tab on graph tab
 
 ### Changed